Create Find_LCM
Java Program to Find LCM of Two Numbers

LCM (i.e. Least Common Multiple) is the largest of the two stated numbers that can be divided by both the given numbers. In this article, we will write a program to find the LCM in Java.

Java Program to Find the LCM of Two Numbers
The easiest approach for finding the LCM is to Check the factors and then find the Union of all factors to get the result.