Create LeapYear_Using_Ternary_Operator
More Methods to Check Leap Year in Java
There are certain methods for leap year programs in Java are mentioned below:

1. Using Scanner Class
2. Using Ternary Operator
3. Using In-built isLeap() Method

2. Using Ternary Operator
Ternary Operator is used to reduce the if-else statements. The ternary operator takes three operands, a boolean condition, an expression to execute if the condition is true, and an expression to execute if false.