Create Vowel_Consonant
For any given character, we need to check if it is a vowel or a consonant. As we know, vowels are ‘a’, ‘e’, ‘i’, ‘o’, ‘u’ and all the other characters (i.e. ‘b’, ‘c’, ‘d’, ‘f’ …..) are consonants.

Examples:  

Input : char = 'r'
Output : Consonant

Input : char = 'e'
Output : Vowel