Create factorial_iterative_Way
Factorial of a non-negative integer, is multiplication of all integers smaller than or equal to n. In this article, we will learn how to write a program for the factorial of a number in Java.

Formulae for Factorial
n! = n * (n-1) * (n-2) * (n-3) * ........ * 1


Example of Factorial in Java
6! == 6*5*4*3*2*1 = 720. 
5! == 5*4*3*2*1 = 120
4! == 4*3*2*1 = 24

Methods to Find Factorial of a Number
1. Iterative Solution for Factorial in Java
2. Factorial in Java Using Recursive Method
3. One-line Solution (Using the Ternary operator)

The complexity of this program:
Time Complexity: O(n)
Auxiliary Space: O(1)