Create Simple_Interest
Simple interest is a quick method of calculating the interest charge on a loan. Simple interest is determined by multiplying the daily interest rate by the principal by the number of days that elapse between payments. 

Simple interest formula is given by:

Simple Interest = (P x T x R)/100 
Where, 

P is the principal amount 
T is the time and 
R is the rate