Create Perimeter_of_a_Rectangle
A Rectangle is a quadrilateral with four right angles (90°). In a rectangle, the opposite sides are equal. A rectangle with all four sides equal is called a Square. A rectangle can also be called as right angled parallelogram.

In the above rectangle, the sides A and C are equal and B and D are equal.

The perimeter of a rectangle is the total length of all its four sides. It can be simply calculated by all its four sides.

Perimeter of rectangle ABCD = A+B+C+D
Since the opposite sides are equal in a rectangle, it can be calculated as the sum of twice of one of its side and twice of its adjacent side.

Perimeter of rectangle ABCD = 2A + 2B = 2(A+B)

Output
The perimeter of the given rectangle of length 10 and breadth 20 = 60
Time complexity: O(1)

Auxiliary space: O(1)