Create Lower Part of the triangle
The pattern has two parts – both mirror reflections of each other. The base of the triangle has to be at the bottom of the imaginary mirror and the tip at the top.

Illustration:

Input:
size = 7

Output:

      * 
     * * 
    * * * 
   * * * * 
  * * * * * 
 * * * * * * 
* * * * * * * 
* * * * * * * 
 * * * * * * 
  * * * * * 
   * * * * 
    * * * 
     * * 
      * 
Approach: 

Divide pattern into sub-patterns

Divide this into two parts
1. Upper part
2. Lower part

Both parts are a mirror reflection of each other. And each part comprises 2 triangles. 

Therefore, In total, we have to print 4 triangles to get the desired pattern.

Output

* * * * * * * 
 * * * * * * 
  * * * * * 
   * * * * 
    * * * 
     * * 
      *