Create Binomial Coefficient
Pascal’s triangle is a triangular array of the binomial coefficients. Write a function that takes an integer value n as input and prints the first n lines of Pascal’s triangle. Following are the first 6 rows of Pascal’s Triangle.In this article, we will look into the process of printing Pascal’s triangle. Pascal’s triangle is a pattern of the triangle which is based on nCr.below is the pictorial representation of Pascal’s triangle.



Illustration:

Input : N = 5
Output:
          1
        1   1
      1   2   1
    1   3   3   1
  1   4   6   4   1
1   5   10   10   5   1

Methods: 
There are two approaches for the same. Let’s check them out.

1.Using nCr formula
2.Using Binomial Coefficient


2.Using Binomial Coefficient

Output
      1 
     1 1 
    1 2 1 
   1 3 3 1 
  1 4 6 4 1 
 1 5 10 10 5 1 

Time complexity: O(n^2) where n is given input for no of rows of pascal triangle

Auxiliary Space: O(1)