Create Octal to Decimal
The octal numbers are numbers with 8 bases and use digits from 0-7. This system is a base 8 number system. The decimal numbers are the numbers with 10 as their base and use digits from 0-9 to represent the decimal number. They also require dots to represent decimal fractions.

We have to convert a number that is in the Octal Number System to the Decimal Number System. The base in an Octal Number is 8, which means that an Octal Number will have digits ranging from 0 to 7.

For Example:
In Octal: 167

In Decimal:(7 * 80) + (6 * 81) +(1 * 82)=119

The below diagram explains how to convert an octal number (123) to an equivalent decimal value:

Java Program For Octal to Decimal Conversion

1. Using Integer.parseInt() method
To convert any string form to decimal, we can use type.parseType() method. For example, here we need to convert from octal to decimal, and the octal form is an integer, so we can use Integer.parseInt() to convert it.

Java Program for Octal to Decimal Number Conversion Using Integer.parseInt() Method

Output
Decimal equivalent of Octal value 157 is: 111
The complexity of the above method:
Time complexity : O(1)
Auxiliary space : O(1)