Create Convert Octal to Decimal
The octal numbers are numbers with 8 bases and use digits from 0-7. This system is a base 8 number system. The decimal numbers are the numbers with 10 as their base and use digits from 0-9 to represent the decimal number. They also require dots to represent decimal fractions.

We have to convert a number that is in the Octal Number System to the Decimal Number System. The base in an Octal Number is 8, which means that an Octal Number will have digits ranging from 0 to 7.

For Example:
In Octal: 167

In Decimal:(7 * 80) + (6 * 81) +(1 * 82)=119

The below diagram explains how to convert an octal number (123) to an equivalent decimal value:

Java Program For Octal to Decimal Conversion

2. Custom Method to Convert Octal to Decimal
Algorithm

1. Start and take the octal input from the user.

2. Create a result variable with an initial value of 0 to store the resultant Decimal number.

3. Create a loop for getting every digit in the Input.

4. Multiply each digit in the number with 8n-1, where n is the digit’s position.

5. Then add it to the result.

6. Store the value in Step(5) to the result variable.

7. Print the result variable.

Output
Decimal of Octal Number (167) : 119
The complexity of the above method:
Time complexity: O(logN) for given octal number N